Making definition of module actions more robust, making home/site assignments
Instead of providing an array of actions an associative array is created
where the keys are the actions and the values are any parameters in the
state's url.

Also, with this new functionality it is possible to separate home and site
interviews into two separate lists when making new assignments.