Fixing bugs
There was a name collision in the appointment module having to do with
qnaire and appointment types (both named type).  The qnaire type is now
named qnaire_type, but the appointment type has been left as simply
"type" since we don't want to update all the Onyx laptops.

There was also a bug in the new feature that adds identifiers to the
machine appointment service.